backend/requirements.txt
django 
djangorestframework 
djangorestframework-simplejwt 
django-filter 
jupyter 
pandas